requirements.txt
boto3
botocore
psycopg2-binary
pyodbc
pytest
